refactor: Parser 클래스 생성
input을 String List로, String List를 Integer List로 바꿔주는 역할 담당
input을 int로 바꿔주는 역할 담당
